requirements.txt
streamlit
st-clerk
requests
Pillow